adding support for image input
the annotate.core module contains the new class ImageInput
to support input in the form of still images. It creates a
batch with a batch size of 1, because in general all of the
images may not have the same size. Not all annotators will
make sense for still images, but should still be supported
whenever possible. Changed some aspects of the batches to
better support image data, most importantly the way that
the set of frames are stored.