fixing bug with kmeans in CIElab
the kmeans function does not always return k clusters
in the case where there are fewer than k unique data
points. This can be a problem when dealing with, for
example, an all black frame. The previous patch fails
when there is more than one instance in a batch. This
fixes that issue.